Fix race condition: stop listening for babeld on stdin
The problem was, that the ctx->babeldfd was initialized to 0 when
we reach the line removed in this commit, so the daemon began to wait
for babeld input on stdin. The ctx->babeldfd is initialized
correctly after the ctx->babeld_reconnect_tfd timeout is reached.

This caused some issues on openWRT/LEDE with procd. Procd does not
open a stdin file descriptor to the daemon, so trying to read on
fd0 is a bad idea.