Temporary fix: add subscription to babeld events
Before this fix, mmfd was connecting to the babeld
socket, but was not subscribing the events using
the "monitor" command.

This commit should not be used in production, since
it does not handle all errors of send() correctly.
If EAGAIN is returned, the application will die.
I think, it should instead try to send() later again.